Share login and register auth entry template
