+++ public_html/src/View/partial/auth-entry.twig
@@ -1,0 +1,60 @@
+{% import "macro/forms.twig" as forms %}
+{% import "macro/alerts.twig" as alerts %}
+
+{% if (uUID or UID) %}
+    <a class="absolute right-0 z-20 m-5 inline-flex items-center gap-2 rounded border border-gray-400 bg-gray-100 px-4 py-2 font-semibold text-gray-700 shadow-sm hover:bg-gray-200 dark:border-gray-600 dark:bg-slate-800 dark:text-gray-200 dark:hover:bg-slate-700" href="{{ docRoot }}logout" aria-label="Logout">
+        {% if authEntry.showLogoutIcon %}
+            <svg class="h-5 w-5" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
+                <path stroke-linecap="round" stroke-linejoin="round" d="M12 3v8" />
+                <path stroke-linecap="round" stroke-linejoin="round" d="M7.05 6.35a8 8 0 1 0 9.9 0" />
+            </svg>
+        {% endif %}
+        <span>Logout</span>
+    </a>
+{% endif %}
+<section class="flex flex-col md:flex-row h-screen">
+    <div class="bg-transparent flex w-full h-full md:w-1/2 xl:w-3/4 items-center justify-center">
+        <div class="bg-logo bg-no-repeat bg-cover w-80 h-80 lg:w-96 lg:h-96"></div>
+    </div>
+    <div class="text-gray-700 dark:text-gray-200 bg-white dark:bg-slate-900 w-full md:max-w-md lg:max-w-full md:mx-auto md:w-1/2 xl:w-1/4 h-screen px-6 lg:px-16 xl:px-12 flex items-center justify-center">
+
+        <div class="w-full h-100">
+            <form class="mt-6 auth-step auth-step-fade" method="POST" autocomplete="off">
+                {% for error in authEntry.flash.errors %}
+                    {{ alerts.error(error) }}
+                {% endfor %}
+                {% for success in authEntry.flash.success %}
+                    {{ alerts.success(success) }}
+                {% endfor %}
+                {% if (uUID or UID) %}
+                    {% if (uUID) %}
+                        <div class="mb-4">
+                            <label for="email_reference" class="block">{{ authEntry.emailLabel }}</label>
+                            <input type="email" id="email_reference" value="{{ email }}" class="auth-input contrast-75 cursor-not-allowed" autocomplete="username" disabled />
+                        </div>
+                    {% endif %}
+                    {% set name, value, type, submitName = 'totp', totp, 'text', 'submit_totp' %}
+                    {% set label = authEntry.verificationCodeLabel %}
+                    {% set submit = authEntry.verifyLabel %}
+                    {{ forms.otpInput(name, value, type, label, value) }}
+                {% else %}
+                    {% if authEntry.mode == 'register' %}
+                        {% set name, value, type = 'username', '', 'text' %}
+                        {{ forms.textInput(name, value, type, authEntry.usernameLabel, authEntry.usernameInvalidText) }}
+                    {% endif %}
+                    {% set name, value, type, submitName = 'email', email, 'email', authEntry.firstStepSubmitName %}
+                    {{ forms.textInput(name, value, type, authEntry.emailLabel, authEntry.emailInvalidText) }}
+                    {% set submit = authEntry.firstStepSubmitLabel %}
+                {% endif %}
+                {{ forms.submitInput(submitName, submit) }}
+
+            </form>
+            {% if not uUID and not UID %}
+                <hr class="my-6 border-gray-400 dark:border-gray-600 w-full">
+                <p class="mt-8">{{ authEntry.switchPrompt }}
+                    <a href="{{ docRoot }}{{ authEntry.switchPath }}" class="text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-300 font-semibold">{{ authEntry.switchLinkLabel }}</a>.
+                </p>
+            {% endif %}
+        </div>
+    </div>
+</section>
