+++ public_html/src/View/register.twig
@@ -1,60 +1,27 @@
 {% extends "tmpl/base.twig" %}
 
-{% import "macro/forms.twig" as forms %}
-{% import "macro/alerts.twig" as alerts %}
-
 {% block md %}
     <title>{{ __('register') }} - GangsterClub.com</title>
 {% endblock %}
 
 {% block content %}
-    {% if (uUID or UID) %}
-        <a class="absolute right-0 z-20 m-5 inline-flex items-center gap-2 rounded border border-gray-400 bg-gray-100 px-4 py-2 font-semibold text-gray-700 shadow-sm hover:bg-gray-200 dark:border-gray-600 dark:bg-slate-800 dark:text-gray-200 dark:hover:bg-slate-700" href="{{ docRoot }}logout" aria-label="Logout">
-            <span>Logout</span>
-        </a>
-    {% endif %}
-    <section class="flex flex-col md:flex-row h-screen">
-        <div class="bg-transparent flex w-full h-full md:w-1/2 xl:w-3/4 items-center justify-center">
-            <div class="bg-logo bg-no-repeat bg-cover w-80 h-80 lg:w-96 lg:h-96"></div>
-        </div>
-        <div class="text-gray-700 dark:text-gray-200 bg-white dark:bg-slate-900 w-full md:max-w-md lg:max-w-full md:mx-auto md:w-1/2 xl:w-1/4 h-screen px-6 lg:px-16 xl:px-12 flex items-center justify-center">
-            <div class="w-full h-100">
-                <form class="mt-6 auth-step auth-step-fade" method="POST" autocomplete="off">
-                    {% for error in register.errors %}
-                        {{ alerts.error(error) }}
-                    {% endfor %}
-                    {% for success in register.success %}
-                        {{ alerts.success(success) }}
-                    {% endfor %}
-                    {% if (uUID or UID) %}
-                        {% if (uUID) %}
-                            <div class="mb-4">
-                                <label for="email_reference" class="block">{{ __('login.email-address') }}</label>
-                                <input type="email" id="email_reference" value="{{ email }}" class="auth-input contrast-75 cursor-not-allowed" autocomplete="username" disabled />
-                            </div>
-                        {% endif %}
-                        {% set name, value, type, submitName = 'totp', totp, 'text', 'submit_totp' %}
-                        {% set label = __('login.verification-code') %}
-                        {% set submit = __('login.verify') %}
-                        {{ forms.otpInput(name, value, type, label, value) }}
-                    {% else %}
-                        {% set name, value, type = 'username', '', 'text' %}
-                        {{ forms.textInput(name, value, type, __('username'), __('provide-valid-username')) }}
-                        {% set name, value, type, submitName = 'email', email, 'email', 'submit_register' %}
-                        {{ forms.textInput(name, value, type, __('login.email-address'), __('login.provide-valid-email-address')) }}
-                        {% set submit = __('register') %}
-                    {% endif %}
-                    {{ forms.submitInput(submitName, submit) }}
-                </form>
-                {% if not uUID and not UID %}
-                    <hr class="my-6 border-gray-400 dark:border-gray-600 w-full">
-                    <p class="mt-8">{{ __('already-have-account') }}
-                        <a href="{{ docRoot }}login" class="text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-300 font-semibold">{{ __('login.sign-in') }}</a>.
-                    </p>
-                {% endif %}
-            </div>
-        </div>
-    </section>
+    {% set authEntry = {
+        mode: 'register',
+        flash: register,
+        showLogoutIcon: false,
+        usernameLabel: __('username'),
+        usernameInvalidText: __('provide-valid-username'),
+        emailLabel: __('login.email-address'),
+        emailInvalidText: __('login.provide-valid-email-address'),
+        verificationCodeLabel: __('login.verification-code'),
+        verifyLabel: __('login.verify'),
+        firstStepSubmitName: 'submit_register',
+        firstStepSubmitLabel: __('register'),
+        switchPrompt: __('already-have-account'),
+        switchPath: 'login',
+        switchLinkLabel: __('login.sign-in')
+    } %}
+    {% include "partial/auth-entry.twig" %}
 {% endblock %}
 
 {% block js %}
